requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
pywebpush==1.14.0

# RAG System Dependencies
faiss-cpu==1.7.4
scikit-learn==1.3.2
google-generativeai==0.3.1
numpy==1.24.3
# Note: Using TF-IDF + Cosine Similarity (lightweight, no CUDA needed)

